requirements.txt
allure-pytest==2.12.0
playwright==1.29.1
pytest==7.2.0
pytest-playwright==0.3.0
python-slugify==7.0.0